Improve readme and request.http, create new route and reformat

+++ fastApi/router/ability_router.py
@@ -1,6 +1,9 @@
-from fastapi import APIRouter
+from typing import Annotated
+from fastapi import APIRouter, Depends
 from controller.ability import ability_controller
 from entity.ability import AbilityEntity
+from service.middleware import get_current_user
+from db.model.models import User
 
 router = APIRouter(
     prefix='/ability',
@@ -24,15 +27,18 @@ async def get_one(name: str):
 
 
 @router.post("")
-async def add_one():
-    return await ability_controller.create_ability()
+async def add_one(current_user: Annotated[User, Depends(get_current_user)]):
+    if current_user:
+        return await ability_controller.create_ability()
 
 
-@router.put("/{id}")
-async def update_one(ability: AbilityEntity):
-    return await ability_controller.update_ability(ability)
+@router.put("/")
+async def update_one(ability: AbilityEntity, current_user: Annotated[User, Depends(get_current_user)]):
+    if current_user:
+        return await ability_controller.update_ability(ability)
 
 
 @router.delete("/{id}")
-async def delete_one(id: int):
-    return await ability_controller.delete_ability(id)
+async def delete_one(id: int, current_user: Annotated[User, Depends(get_current_user)]):
+    if current_user:
+        return await ability_controller.delete_ability(id)
