+++ fastApi/router/pokemon_router.py
@@ -1,6 +1,11 @@
-from fastapi import APIRouter
+from typing import Annotated
+from typing import Annotated
+from fastapi import APIRouter, Depends
 from controller.pokemon import pokemon_controller
 from entity.pokemon import PokemonEntity
+from service.middleware import get_current_user
+from db.model.models import User
+
 
 router = APIRouter(
     prefix='/pokemon',
@@ -24,15 +29,18 @@ async def get_one_name(name: str):
 
 
 @router.post("")
-async def add_one(pokemon: PokemonEntity):
-    return await pokemon_controller.create_pokemon(pokemon)
+async def add_one(pokemon: PokemonEntity, current_user: Annotated[User, Depends(get_current_user)]):
+    if current_user:
+        return await pokemon_controller.create_pokemon(pokemon)
 
 
 @router.put("")
-async def update_one(pokemon: PokemonEntity):
-    return await pokemon_controller.update_pokemon(pokemon)
+async def update_one(pokemon: PokemonEntity, current_user: Annotated[User, Depends(get_current_user)]):
+    if current_user:
+        return await pokemon_controller.update_pokemon(pokemon)
 
 
 @router.delete("/{id}")
-async def delete_one(id):
-    return await pokemon_controller.delete_pokemon(id)
+async def delete_one(id, current_user: Annotated[User, Depends(get_current_user)]):
+    if current_user:
+        return await pokemon_controller.delete_pokemon(id)
